refactor: split action brief reports
